requirements.txt
streamlit
pandas
astrapy
langchain
openai
python-dotenv
langchain_community
pandas
transformers
torch
google-generativeai